Removed agent ID specification from the threat event create example
Previously, the threat event create example required the user to
determine an agent ID to use in the threat event payload. Since the DXL
ePO extension can now determine the agent ID automatically from the IP
or MAC address under the 'source' object in the event payload, the
example has been updated to only set the IP, MAC, and hostname source
(removing the logic to set the agent ID).